reference: document possibility of spurious CAS failures
An operation like `box-cas!` can fail spuriously on some platforms,
such as ARM processors. Mention that in the documentation.

The `box-cas!` and similar implementations could instead retry on
spurious failure, and that would avoid potential problems with code
tested on x86, where there's no spurious failure. In this case,
though, it seems better to stick with the existing behavior and
stay closer to the machine operation.