Add arrows from definitions to all-defined-out
Currently, `(all-defined-out)` expands to all identifiers defined in a
module. However, this won't cause arrows to be drawn from definitions
to `all-defined-out` because the source location of expanded identifiers
are of definitions rather than `(all-defined-out)`.

This PR fixes the source location of the exported identifiers.
However, since these identifiers are potentially armed, modifying it
will cause it to be tainted and unusable. The workaround is to
use the syntax property `disappeared-use` instead, and since we are
modifying the syntax, we also need to mark `original-for-check-syntax`.

Other provide transformers whose `export-orig-stx` is faithful could also
benefit from this PR. Provide transformers like `struct-out` however is
not faithful (its `export-orig-stx` is struct definitions rather than
`struct-out` itself), so no arrow will be drawn.