reqprov: fix disappeared-use
This PR fixes several issues regarding disappeared-use in reqprov.rkt

1. Add `disappeared-use` for each exported identifier expanded from
`all-defined-out` with a syntax location of `all-defined-out` so that
Check Syntax can draw arrows from definitions to `all-defined-out`.

2. Copy/preserve disappeared-use for all provide subforms instead
of only the first one. The following program doesn't have an arrow from
`struct-out` to `#lang racket` prior this PR, but it does after this PR.

    #lang racket
    (provide a? (struct-out a))
    (struct a ())

3. Fix incorrect `disappeared-use` for `struct-out`: it is inappropriate
to check if an identifier is bound to a struct info in provide
pre-transformer because it would not be able to handle backlinks.
This PR moves the attachment from provide pre-transformer to provide
transformer, allowing Check Syntax to draw an arrow for the identifier
`abc` in:

   #lang racket
   (provide (struct-out abc))
   (struct abc ())