expander: repair literal syntax in test position of `if`
Fix mishandling of an expanded `if` where the test position is a
syntax object. The expander's compiler pass from expanded objects to
linkets knows that the syntax object isn't useful, but it tried to be
helpful by preserving the syntax object's content as quoted --- and
that content turns out not to be available, so the syntax object was
replaced by #f, instead.

Closes #3436

(cherry picked from commit 1a1bad4e90b07b672f2cc7de47d808790e71d576)