Add sequence conversion functions for built-in collection types
This commit adds the following functions:

- sequence->bytes
- sequence->string
- sequence->vector
- sequence->set

Each of these functions is equivalent to the corresponding list-specific version, such as list->vector, except they accept arbitrary single-element sequences instead of lists.  The functions include two important optimizations: fast paths for specific kinds of sequences and a fast path for already-converted inputs.

The functions are meant to serve as drop-in replacements for their monomorphic equivalents, with the exception that they always produce immutable values instead of mutable ones. To simplify the APIs of these collection types, all other forms of converting between these collections are soft-deprecated in favor of the unified sequence conversion APIs.