raco-pkg: show full path instead of broken front page
`raco pkg` currently shows a message like:

Resolving "pollen" via https://download.racket-lang.org/releases/7.9/catalog/

However, this URL is not meant to be visited, giving 404 to visitors,
causing them to think that the package server is down, even though
that might not actually be the case.

The problem is exacerbated by the now-fixed #3499, but
I have seen this problem even before the 7.9 release (at least 5 times),
so it seems worth fixing.

In general, we should not show 404 URLs to users, anyway.

Note that I didn't change the local database and local directory cases
because I don't know how to best display the information.
However, since they are local, users at least won't (incorrectly) think that
the server is down.