cs,compiler/embed: write code for target machine when cross-compiling
Racket CS cannot currently read fasl files for platforms other than
the host, but `compiler/embed` has to be able to read compiled code in
order to figure out what code needs to be embedded into an output
image and which runtime paths need to be included.  This change makes
it so that host code is used to figure all of that information out,
but that code is then replaced by target machine code before it is
written to the output image.

The new logic only applies when the right cross-compilation flags are
set (per `cross-multi-compile?`).